Cloud Functions/Vertex AI - Batch prediction/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-bigquery
google-cloud-aiplatform